Moved devmap intialization from LanItems to interface builder
